Set docker-compose file version to the minimum required
We noticed an issue after redeployment where the version specified in the
docker-compose.yml file was not supported, and docker-compose would not run the
compose configuration. This was remedied in production by setting the version
to '3.3'. After discussion with @jsf9k, I am setting these compose files to the
minimum docker-compose version they require.

The
- docker-compose.bod.yml
- docker-compose.cyhy-notification.yml
- docker-compose.cyhy.yml
files use long syntax volume mounts so they require version '3.2', and
- docker-compose.yml
uses the secrets parameter, so it only requires version '3.1'.