Remove the redundant image key from docker-compose.*.yml files
In PR #84 I updated the cyhy-mailer Docker image version in
docker-compose.yml, but the change did not take.  I eventually figured
out that this was because the image key (which was pointing to the
previous version of the image) also exists in
docker-compose.{bod,cyhy-notification,cyhy}.yml.  To make it so that
we only need update the image in a single place, I am removing the
image key from docker-compose.{bod,cyhy-notification,cyhy}.yml.