Replace some uses of the `ncats-dev@beta.dhs.gov` email address
We have migrated from the legacy Google Workspace account that houses
this email so we should update to a group that resides in the Google
Workspace Enterprise account that we now use. This updates the Docker
compositions in the repository to use a new email address in the GWE
account.